chore, extend quo_s index field to 32bits (like sub's idx slot)

+++ src/quo.h
@@ -63,7 +63,7 @@ typedef enum {
 } quo_typ_t;
 
 struct quo_s {
-	uint16_t idx;
+	uint32_t idx;
 	quo_typ_t typ;
 	double val;
 };
